requirement.txt
Flask
torch
torchvision
pillow
opencv-python
numpy
